+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # CHANGELOG
 
+# 0.1.73
+
+updated to 0.1.73, here are all the changes
+- added CustomElement function for custom html elements
+- added new paramter to application config (allowzoom)
+- added onScroll method to Application class.
+- exported all types
+- layouts dont take all space available
+- updated docs
+
 ## 0.1.7
 
 updated to 0.1.7, here are all the changes